fix loc

+++ src/main/java/com/wild/daos/IHistoryDao.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.wild.models.History;
 
 public interface IHistoryDao extends GenericDAO<History>{
-		List<History> getListHistoryByIdUser(int idUser);
+		List<History> getListHistoryByIdUser(Long idUser);
 }
